Set yum options to fail when one package in the list does not exist
Summary:
Yum from CentOS 7 has a backport that allows setting config options to enable this behavior of faling when one package in a list does not exist. Enabling these options will allow us to detect when a package has been requested, but couldn't be installed.

This is standard behavior in Dnf, so now one of our tests cases that had to branch based on Yum or Dnf can be consolidated, since both now match in this behavior.

I also added a test to install a package by one of its "Provides:", using `install-n`, since that's what we actually use in `image.rpms_install()`. Unfortunately here the behavior from Yum and Dnf diverges again, with Dnf expanding the "Provides:" of the package even in `install-n` while Yum doesn't and fails to install it. I added a test to assert this behavior on both.

Reviewed By: snarkmaster

Differential Revision: D20450916

fbshipit-source-id: b5b7e7062d77ec58d36c4a1e3783abbbdc541f63