Specify repo server ports as part of the RPM snapshot
Summary:
Since `dnf` treats the repo URLs as part of the cache key, we need to hardcode the `localhost:port` combo for the `repo-server`. This also simplifies the injection of snapshots into containers.

In the unlikely event we hit a build-time port conflict, the ports can be changed per-snapshot.

I'm specifying multiple ports because spinning up multiple `repo-servers` can give a significant download speedup in some contexts. No matter the configuration, `yum` will not do parallel downloads against a single repo URL.

Note that this diff is a no-op, it just adds some minimal plumbing. The context is that a later diff will use this to save rewritten `yum/dnf.conf` files as part of building the snapshot. And the next step after that is to use the pre-built configs in `yum-from-snapshot`, simplifying its flow.

Reviewed By: zeroxoneb

Differential Revision: D20451908

fbshipit-source-id: 08a22e1f29f9b09e2b26677f4ab42d99f4b74ba9