Move `qemu_bioses` into a `/facebook` directory
Summary:
### Goal (Copy-Pasta'd from D20517601)
After a discussion with vmagro, it turns out that we only need the list of BIOSes for FB-specific reasons. In the OSS installation, these BIOSes should be present with the installation of QEMU. As such, it would be nice if we could move the BIOSes to a /facebook gated directory. The qemu_bioses are being loaded as a resource and it seems like /facebook normalization only occurs for deps atm.

### Changes
- Move `qemu_bioses` to `//fs_image/facebook/vm` and update references
- Add a try-except clause when loading the `qemu_bioses` in `vm.py`. If a FileNotFoundError is raised, `kernel.qemu_bioses` will be `None`.
- Only add the `-L <BIOS-path>` argument if `kernel.qemu_bioses` is not `None`.

Reviewed By: snarkmaster

Differential Revision: D20506797

fbshipit-source-id: 3c115fb44a770f4720783a50f12771e747ced34e