`Path` & `argparse` improvements
Summary:
## Rationale

The direct motivation for this diff is that during development, we lose a fair bit of time figuring out whether to call `Path.decode()` when interfacing with other code.

## Changes

This diff eliminates about 45 callsites of `.decode()`, and provides clear patterns for handling the common problems of `argparse.parse_args` and `json.dump` not accepting bytes.

The only change to business logic is the addition new `Path` methods, which I am happy to break out into a separate diff if reviewers ask.

The rest of the diff uses these to make cleanups as appropriate:
 - Introduce `Path.parse_args` and `Path.json_dump{,s}`.
 - Use `Path.listdir` instead of `os.listdir`.
 - Normalize `--long-arg` handling to use the `=` form using the newly added `Path.__format__`.

## New subprocess argument convention

Use the `--foo=bar` form for long-form args, instead of separate arguments, unless:
   - The target command does not accept the = form, and requires two separate arguments.
   - The option takes multiple arguments. Then, the suggested form is:

```
subprocess.Popen([
    '--one-arg=foo',
    # Group to prevent `black` from putting this on 3 lines
    *('--multi-arg', 'bar', baz'),
])
```

Reviewed By: filbranden

Differential Revision: D20544158

fbshipit-source-id: c467f37b0b4a5720775c59a21a35228707efc087