Use pre-built `{yum,dnf}.conf` to serve repo snapshots
Summary:
Instead of generating `yum` / `dnf` configs on the fly, we now use configs deposited into the snapshot build artifact by `rpm/write_yum_dnf_conf.py` (D20457517).

This lets us delete swaths of now-unused code in `nspawn_in_subvol/repo_server.py` and `yum_dnf_from_snapshot.py`.

Note that, as of now, containers can't yet populate  `versionlock.list` (that's still locked up in `yum_dnf_from_snapshot.py`), but it's a tiny step away now. I could have made `versionlock.list` be part of the snapshot build artifact, but that's not how we want to use version pinning --- we don't want to be rebuilding the build appliance to update the version pin list. So, instead, the plan is to inject a `versionlock.list` at runtime.

As an incidental refactor, this normalizes the naming of the option to serve a RPM snapshot in a container. Now it matches the regex `serve.rpm.snapshot(|s)` throughout.

Reviewed By: zeroxoneb

Differential Revision: D20462816

fbshipit-source-id: 5f380735cabd4530fd271cbb97b869c012bc730c