`nspawn_in_subvol` allows wrapping the `popen` impl used by the `run` variants
Summary:
Serving RPM repo snapshots isn't a core feature of `nspawn_in_subvol`, so it's problematic for it to be integrated into the guts of `non_booted.py`.

It's cleaner to add such features via a wrapper, as this allows us to inject multiple proxies into containers with ease, without one interfering with the other.

The only thing that we need to wrap is `popen`, first changing its `opts`, and later interacting with the started container.

This here makes `run` support wrapping `popen`, for the common case where people want to use the `run` API.

Reviewed By: zeroxoneb

Differential Revision: D20573180

fbshipit-source-id: aca79d8e37e6ded999548a63f333de42012014a0