Cleaner split of responsibilities at the `nspawn_in_subvol` CLI layer
Summary:
The rationale is that we had roughly the same plumbing from `parse_cli_args` into `run_nspawn` in 3 places, and with the addition of `popen_wrappers`, the amount of complexity that was not explicitly covered by the module's tests was getting a bit high.

This forces all 3 callsites to use a single data-flow:
 - `PopenArgs` is specified at the callsite, except for `boot_console`
 - `_CliSetup` represents the rest of the dataflow

This is reasonably clear, and is now explicitly covered in `test_run.py`.

Reviewed By: zeroxoneb

Differential Revision: D20630715

fbshipit-source-id: 315213e9a2365e3110a719e7f34647f11419a745