`Path.resource` for correct resource access
Summary:
This should let us eliminate `load_location` / `target_location`, as well as many instances of `par_style` in the codebase.

The `*_location` pattern is especially bad since it injects a repo- and host-specific path into the built binary. If such a binary makes it into Buck's distributed caches, it will break on any clients that fetch it instead of building it.

Having a robust solution to the resource problem is important because:
(a) we keep breaking mode/opt,
(b) we keep breaking open-source PEX support.
This new new `Path.resource` should handle both of these cases gracefully.

The only place where I know we still need `__file__` is to refer to layers from tests. The problem there is that such resources are **directories**, not **files**, and this is not something that `importlib` supports. This can be fixed, I have a comment on this inside a stacked diff.

Reviewed By: zeroxoneb

Differential Revision: D20630327

fbshipit-source-id: b93c79569fb7f9a697641626b05a5ba5a3e60a7a