Let `TempSubvolumes` clean up undeclared inner subvols
Summary:
Booting a `systemd` image tends to create subvolumes at `/var/lib/{machine,portable}`, which would break the cleanup of the temporary subvolume.

In the past, we would generally disable these systemd features by changing the in-image configuration, but this is cumbersome / distracting, so it's better to make the cleanup more robust.

The code is horrible because:
 - `btrfs subvol list` is very, very broken -- this here is the only regime I believe to work correctly
 - using `libbtrfsutil` would be nicer, but involves extra work (a) creating a small binary so we can `sudo` it, (b) making the OSS build include `libbtrfsutil`. I'm game to switch to it if (b) ever happens.

Reviewed By: zeroxoneb

Differential Revision: D20640234

fbshipit-source-id: eab752c361ab503b12e5a23d61d6e0b83e11c6a7