`nspawn_in_subvol` gets repo servers via a wrapper, works in booted
Summary:
As described on D20573180, it was not scalable to tightly couple the serving of RPM snapshots to the internals of `non_booted.py`.

The new layout makes it so that only the very top level of the CLI knows anything about repo snapshots (for human convenience).

And in-Python customers of the `nspawn_in_subvol` CLI can apply repo-serving as a wrapper, such that it composes cleanly with any other wrappers they may want.

As a bonus, repo serving now works in booted containers, too.

Reviewed By: zeroxoneb

Differential Revision: D20640233

fbshipit-source-id: 8088654ac564eb105f5de6a8e4d918e049353c94