Temporarily comment test verifying that by-version install fails
Summary:
A recent `dnf` update changed the behavior of `install-n` to accept a `name-version`, so this specific test fails.

As it turns out, we were only running it against `dnf`, so commenting it out is the easiest for now. In the not-too-distant future, we'll validate package names on our own, which will make moot the subtleties of `dnf` and `yum` behavior.

(There's some similar coverage as part of `test-yum-dnf-from-snapshot`.)

Reviewed By: filbranden

Differential Revision: D20725174

fbshipit-source-id: 1407c84ade6dbc344ab895f264f9ea067d03b72f