Make `nspawn_in_subvol` opts debug-only: `--no-private-network`, `--cap-net-admin`
Summary:
These used to be used by `RpmActionItem` because it used to start the `repo-server` inside the build appliance. Now that we run it externally, the BA container doesn't need network access, and we don't need `--cap-net-admin` because `yum-dnf-from-snapshot` no longer needs to do `unshare --net` (for the same reasons).

The point of marking these debug-only is to make sure we think really hard before shipping any "production" feature that uses these.

There's some case (nested containers) to be made for `--cap-net-admin`, but `--no-private-network` should probably never be exposed again.

Reviewed By: filbranden

Differential Revision: D20797117

fbshipit-source-id: 6af7db011d399704215b3f2f017787f235b04bee