Remove base_module from fs_image/rpm
Summary:
Was in the mood to grind out some busy work this morning and figured it was easier to just rip off the bandaid here than continue to add more stuff using `base_module`. As part of this I also fixed the weird pattern for common imports we were doing, hiding where it was actually coming from.

More to come for the rest of `fs_image` but figured this was a good starting point.

Note once we're able to autoformat this directory that should handle sorting these imports for us, as they're now likely out of order from this.

Reviewed By: filbranden, snarkmaster

Differential Revision: D20820960

fbshipit-source-id: b0f45d70c869a4e1ba2daf3ecf110b59e0dbe68c