Inject `versionlock.list` using an nspawn wrapper
Summary:
One high-level goal is to shrink `yum-dnf-from-snapshot` as much as possible, so that deviations from running vanilla `yum` / `dnf` in the container are minimal.

Also, it is not the right place to inject version-locking information since we might want to version-lock for EVERY user of a snapshot of the container, not just those who know to pass the right arguments to `yum-dnf-from-snapshot`.

Lastly, using the `nspawn_in_subvol` wrapper API makes this be rather loosely coupled with all the other parts of the system, as it should be.

Reviewed By: zeroxoneb

Differential Revision: D20827661

fbshipit-source-id: fdb2ee947c009c590eb438ac177d257426f498d9