vmtest: replace seedroot with `btrfs device add` in init.sh
Summary:
`seedroot` was a statically compiled binary to setup the r/w root mount.

This required `glibc-static` to be installed on the host machine, which is
apparently not true everywhere. This diff replaces it by calling `/sbin/btrfs`
that is installed in the root image to do the same thing, without needing
`btrfs` to exist in the initrd.

Reviewed By: snarkmaster

Differential Revision: D19828172

fbshipit-source-id: a95b5e6b0cad63ccc0ce17646093f8f816a02ac7