Implement `image_rpmbuild_layer` in terms of `image_foreign_layer`
Summary:
Also move all `rpmbuild`-related stuff into `bzl/foreign`, as per `image_foreign_layer` rules.

Note that we no longer have a compiler item for this, that entire logic got replaced by a much simpler `cmd = [...]` kwarg in the `.bzl` of the macro.

And the tests are now in one place, and pretty straightforward.

Moreover, I made the "install" test use subvol rendering, just because it was easy, and asserts the entire state of the filesystem.

Reviewed By: zeroxoneb

Differential Revision: D20901171

fbshipit-source-id: 1a7a15a17ef86efdf437882e389862bc73065d54