More uniform `outerproc` handling
Summary:
As written, `inject_repo_servers.py` would not work as part of `foreign_layer.py`, because that write-protects `__fs_image__` to prevent e.g. the RPM snapshots from being corrupted.

Regardless of all this, we will probably want to provide an RW area inside of `__fs_image__` to ease cleanup after foreign layer artifacts. I'll do that when I have a clear use-case.

However, for `outerproc`, I realized that we have 2 copies of this code, and the one in `booted.py` **has** to be at `/` since `/__fs_image__` isn't guaranteed to be present. At that point I decided that we might as well just do it uniformly.

I gave the two outerprocs unique names because otherwise the `unmount` / `rm` cleanup would fail when both are in use.

Reviewed By: zeroxoneb

Differential Revision: D20930438

fbshipit-source-id: 57d89a7d8472a5833167ce0fab43c73b06cb330f