Make our `yum-dnf-from-snapshot` wrappers transparent
Summary:
My theory is that the little bit of extra filesystem isolation / sanity checking that we do in `yum_dnf_from_snapshot.py` is overall valuable to the determinism & correctness of operating on RPM snapshots.

If so, we should **always** use this binary for container RPM installation (as opposed to `dnf -c /__fs_image__/...`). To achieve this, we need roughly 3 things:
 - That our snapshots' `bin/{yum,dnf}` scripts be drop-in replacements for the original.
 - An `nspawn_in_subvol` wrapper that uses some bind-mount trickery to shadow the underlying OS's `yum` and/or `dnf` with the snapshot variant.
 - Optionally, we can "corrupt" our configs so they can't be used with `dnf --config=` out of the box.

This diff does step 1 of the program.

Reviewed By: zeroxoneb

Differential Revision: D20958002

fbshipit-source-id: 635842d2ec7f661d0e639d7ce0fb573c36164039