Change some `.bzl` comments into docstrings
Summary:
This makes some small improvements to the content of the docs.

The comment -> docstring change was prompted by a Buildifier docstring linter that was automatically enabled by accident. I have since disabled the noisy lint, but was too lazy to reformat this as comments, since I have no strong preference.

Reviewed By: justintrudell

Differential Revision: D20987897

fbshipit-source-id: 4944fe21c9acf4de7fa93e9a25ed6b4663a74646