Fail fast on vm failure
Summary:
Currently we rely on the QEMU guest agent (GA) for checking inner test liveness. As a result, when the vm fails due to kernel panic or other halt events, the GA won't respond to our liveness poll and we hang until timeout. It is desirable to fail fast on those cases.

I tried with the QEMU Machine Protocol (QMP), which can query vm level info. However, it does not detect kernel panic correctly.

From this post https://unix.stackexchange.com/questions/443017/can-i-make-qemu-exit-with-failure-on-kernel-panic someone suggested using the `pvpanic` device, which requires kernel support so I didn't tried.

This diff adopts the other option from the same post that reboots kernel on panic and stop QEMU on reboot.

Since now QEMU may stop in the middle of the test, we need to handle exceptions due to GA connection lost properly. Set default return code to -1 (255) to indicate that QEMU exits unexpectedly.

Reviewed By: vmagro

Differential Revision: D20984707

fbshipit-source-id: 08dc0edab7107208caf4d04dbe287c6b42c2cbba