vmtest: mount modules over 9p share
Summary:
As lnyng pointed out, the size of the initrd can cause timeouts if the
Sandcastle host is under memory pressure. The initrd is huge because all kernel
modules are included to avoid per-kernel images.

This diff mounts all kernel modules in a 9p share, and leaves only the bare
minimum modules in the initrd (virtio_blk, 9pnet, 9pnet_virtio, 9p).

Reviewed By: zeroxoneb

Differential Revision: D20995777

fbshipit-source-id: a1ceedab9e4d806efa4822d764889cd47e3b71bc