`__fs_image__/rpm-repo-snapshot` -> `__fs_image__/rpm/repo-snapshot`
Summary:
NB: chef-solo tests will be red until D21246633 lands, and I update the pinned opsfiles hash.

However, this is fine to review, since I checked those tests locally (as per D21246633 test plan).

This is the first in a fairly large reshuffle of how the snapshot directories are laid out.

The overall point of this sequence of renames is to (a) change the defaulting semantics for RPM installers & snapshots -- this is justified on a later diff, (b) make it more obvious how to add other package formats & installers (zeroxoneb has one in mind).

Reviewed By: filbranden, zeroxoneb

Differential Revision: D21246361

fbshipit-source-id: 9cfa052e0ab31ff8fdc5a652b129b818915440b8