Fix flaky stress test
Summary: A new test that was introduced as part of `image.clone` is flaky under stress testing.  This is due to the same src subvol being used concurrently and being marked as `rw` after being rendered as a sendstream.  To fix this, I added a flag to the `render_subvol` test helper to not mark the sendstream as `rw`.

Reviewed By: justintrudell

Differential Revision: D21297610

fbshipit-source-id: 3befbebbe2d522d01b042fa4c845b3631a5053ab