Change RPM repo snapshot & installer defaulting behavior
Summary:
The high-level rationale for this is that I want to shadow the system `dnf` and `yum` binaries transparently in `buck run :layer-container` runs, and this requires the image to know which snapshot to serve for which RPM installer.

Before:
 - Images have a notion of a default snapshot
 - Snapshots know the default RPM installer
After:
 - Images know a default snapshot or each installer, in `/__fs_image__/rpm/default-snapshot-for-installer/{yum,dnf}`.
 - The default RPM installer is chosen via `.buckconfig`

This rearranges all the plumbing to work accordingly, and slightly refactors the RPM-snapshot-install macros to accommodate the new reality.

There are a couple of small drive-by cleanups:
 - The constant `BUILD_APPLIANCE_TARGET` became `DEFAULT_BUILD_APPLIANCE`.
 - The snapshot argument `yum_dnf` became `rpm_installers` (no longer order-sensitive since we don't have a notion of a default installer for a snapshot).

Reviewed By: zeroxoneb

Differential Revision: D21310391

fbshipit-source-id: d94f50fbaa9ab019be9b29fac819171cab11426d