Make `PATH` handling identical for booted & non-booted cases
Summary:
I noticed that we had no contract around `PATH`, and that it was different between the booted and non-booted code paths.

I needed there to be some contract so that I can sanely shadow system binaries, so I standardized the behavior on "what `systemd-nspawn` does by default" for now, and added a test.

Reviewed By: zeroxoneb

Differential Revision: D21333149

fbshipit-source-id: 7e7ff3e3572fb60a1a4de831c8ca6961bcf7dc5e