sign test RPMs and repos with a test key; enable gpgchecks for tests
Summary:
First step to enabling `gpgcheck=1` and `localpkg_gpgcheck=1` for all RPM installations is to make sure all test RPMs are signed and the correct public key is added to the yum/dnf configs for verification.

For `temp_snapshot.py` and the already built `rpm-test-cheese*.rpm`, we use a new target, `generate-gpg-test-keypair` to generate a GPG key pair for the run for signing and verification.

`rpm-test-cheese*.rpm` needed some additional tweaks to get the key installed as mentioned in the comment in `fs_image/compiler/test_images/TARGETS`. I also had to rebuild it to x86_64 (instead of noarch) so that it would match and replace the RPM from `non-default-repo-snapshot-for-tests` during installation (which default to x86_64). If the architectures were mismatched (x86_64 vs noarch) it would not detect that they were the same RPMs.

Reviewed By: snarkmaster

Differential Revision: D21039449

fbshipit-source-id: 2566ad6fef023215fe996e3ae3062be0ce492e1d