Use a containerized build environment for `librename_shadowed.so`
Summary:
We don't want to be building this `.so` using the outer Buck, because its C toolchain (compiler, libc) may be significantly different than what runs in the container.

So instead, let's build in the same environment that will perform the `LD_PRELOAD`, which is the Actually Correct (TM) approach.

Note that this diff also makes the `LD_PRELOAD` code fails open if the `.so` is not available in the image. This is completely desirable because most images will not be wanting to upgrade their package manager in-place. In fact, in-place upgrades are an anti-pattern since it is harder to reason about -- the result of other package installs can depend on the timing of when the package manager gets upgraded.

Reviewed By: zeroxoneb

Differential Revision: D21443711

fbshipit-source-id: 4e922475df53daa045f0383f84c5103ff769df7f