Allow `yum` & `dnf` to upgrade shadowed files
Summary:
We're transparently wrapping `yum` and `dnf`, and the wrapper is read-only bind-mounted on top of their locations (normally in `/usr/bin/...`). This is great -- our wrappers correctly install from our RPM repo snapshot, and this behavior is zero-configuration from the point of view of any user of the container.

However, we need to support `yum install yum` and similar upgrades of the package manager by the package manager. This is a fairly crazy requirement, but it has to do with the following sequence of interesting choices by other parts of the ecosystem.
 - Chef doesn't mind changing the `yum` / `dnf` version that it uses mid-run.
 - Chef hardcodes the location of `yum` / `dnf`, and ignores `PATH`.
 - Certain image builds want to upgrade `yum` / `dnf` relative to a pre-installed one.

The fix in this diff is "surgical" in nature:
 - We interpose on `rename` via `glibc`, which is the only correct way for a package manager to update an existing file (caveat: you could call the syscall directly, but we're in luck). My tests show that this interposition indeed works for both `yum` and `dnf`.
 - The interposed `rename` checks if a path that would be overwritten is shadowed, and changes the destination path to overwrite the shadowed original instead of trying to overwrite a read-only bind-mount.

The main risk of this approach is that at a future point, `yum` or `dnf` will change to a different `glibc` call, like `renameat*`, or (less plausibly) `link`, similar. That would cost us extra implementation effort on updating the shim, but should otherwise be uneventful. `LD_PRELOAD` also has some failure modes that we're not too worried about:
  - Callers making the syscall directly, without a `libc`. Both `yum` and `dnf` currently use `libc`-based toolchains.
  - `LD_PRELOAD` failing to take effect due to being excluded by `secure_getenv` -- we run everything as `root`, so this shouldn't happen.

Note that these approaches don't really work:
 - Anything involving symlinks is racy, since `/usr/bin/yum` would at least transiently become un-shadowed after an upgrade.
 - `PATH` or other manipulation of **which** `yum` is used is impossible because of how Chef works.
 - One would be tempted to just unmount the shadow from the `unshare --mount` that we already have inside `yum_dnf_from_snapshot.py`, but this runs afoul of a fundamental property of how mounts work in Linux. Specifically, if, while inside the private mount NS, we unmount enough stuff that `rename('new yum', '/usr/bin/yum')` succeeds, then the inode that is the mountpoint for the "shadow" bind-mount will be deleted globally (across all namespaces), and so the parent mount NS will automatically remove the shadow bind-mount. The net effect is that this behaves just like a symlink -- it's racy because the moment that an upgrade succeeds, the `yum` binary is no longer shadowed.

Using BPF-based interposition would be more bulletproof, but it comes with a higher implementation cost, so we don't bother.

The worst part about this diff is the compilation setup for the `LD_PRELOAD` library. I will be stacking a separate diff to move the build into the BA.

Reviewed By: zeroxoneb

Differential Revision: D21390244

fbshipit-source-id: 9b23a424606212873daa9365bccb1066642db611