Fix btrfs_diff test to not use module level global for counter
Summary:
Globals are bad, they break in weird ways.  In this case, this breaks because the default test runner in OpenSource Buck runs all test cases in the same process.  This has the effect of incrementing a global counter more times than is expected.  This particular counter is used in 2 separate test cases and the 2nd one executed in OSS would always fail.

Instead, lets use a range iterator defined in the function where the incrementing is actually occuring.

Reviewed By: snarkmaster

Differential Revision: D21701578

fbshipit-source-id: f0787496f182c6e033572fbdca8e38118ffce941