Clean up lifecycle management
Summary:
Sorry for the fairly large diff, but the stacking was beginning to get a bit counterproductive and I'd prefer to be able to do all of the churn on this diff prior to the fbpkg stuff. Also apologies for the changes from recent reviews but some earlier comments by Alexey gave me some inspiration to clean up the lifecycle management here.

The overall idea is to make it more of a "batteries included" approach rather than allowing callers to fine-tune parameters, as this shouldn't be necessary in the large majority of cases.

Mainly, I leverage context managers and two separate classes to get the "isolated daemon" and "single-use" semantics now, which I think ends up with a cleaner overall reading experience. I've also gone ahead and implemented a concrete location where both the backing and checkout will be placed under `buck-image-out`, in line with what was mentioned on D21331735. In addition, I added the option to instead specify `checkout_name` to have a named checkout within that dir. Finally, I made a `transient_checkout` decorator to give us the "isolated checkout" behaviour we had been going for in tests.

What this offers is a clean entry-point for tests to call into to get an Eden checkout, while also still being mildly explicit about the transiency.

I also added a considerable amount of comments throughout to try to clear things up and help potential future consumers.

Reviewed By: snarkmaster

Differential Revision: D21630347

fbshipit-source-id: ecfe0ce0afe517fe955ed5e40e811ad2b7590ae9