Migrate generated non-`.py` sources from `srcs` to `resources`
Summary:
This is mostly a noop for regular builds.  However, generated
sources which aren't really Python sources in the `srcs` add
unnecessary overhead to e.g. `#source-db` support (e.g. in
the form of generating the source, even though it's not used).

Move these to the `resources` parameter to fix this (when
D21804847 lands, it'll make this an error going forward).

Reviewed By: igorsugak

Differential Revision: D21804217

fbshipit-source-id: 163ca242eec42f7db5c48a84ca44cc8bc9d163fb