enhance sudo failure handling in fs_image/unshare.py
Summary:
Due to a faulty sudo, I had occasion to see what happens when commands
in unshare.py fail.  First the "assert nspid[0] == ..." line fails because
index 0 is out of bounds for an empty list.  Then os.close(self._root_fd) fails
because self._root_fd is null.  This diff attempts to clean up the error
handling a bit to focus on the main culprit, whatever it is unexpected about
sudo that leads to "sudo: unable to fork: Cannot allocate memory".

Reviewed By: zeroxoneb

Differential Revision: D21844785

fbshipit-source-id: 55395dec54fa4a12ef5621ab150efcbe88e52e17