s/target_location/layer_resource/
Summary:
We don't want to use this for anything but layers due to the cacheability issues described in the docs on this diff.

OTOH, layers are inherently uncacheable anyway (at least until Buck gives us the tools to fix this), so using this for layers does not introduce any new problems.

Reviewed By: zeroxoneb

Differential Revision: D21914649

fbshipit-source-id: 2446a5e077cb16e3e43612871afa7a3b65dedbf8