s/load_location/layer_resource_subvol/
Summary:
Underneath the hood this is still the original `load_location` counterpart to the original `target_location`.

However, this streamlines the callsites, and prevents accidental uses of this gadget for non-layer resources. By putting the import under `fs_image.tests.`, I'm also trying to signal that using this outside of tests is a bad idea.

Reviewed By: zeroxoneb

Differential Revision: D21918902

fbshipit-source-id: 054f06d17ab80f42e76e5086cada2f732bbb663a