nspawn: `s/wrapper/plugin/`; make `run`/`popen` interface the same
Summary:
Besides a straight-up rename, this adds a `plugins` argument to the `popen_` functions. These are currently not used by customers, so both are no-op changes.

The goal in making the `run_/popen_` interface uniform with respect to `plugins` is so that the user experience of plugins is completely decoupled from their implementation details.

Reviewed By: zeroxoneb

Differential Revision: D22082938

fbshipit-source-id: 92155f175de21c453f1da9a6fe1cfb38fb0963e3