Break out `nspawn_in_subvol` plugin tests from `test-run`
Summary:
This is mostly just a code move, so that future plugins don't have the temptation to pile more tests into the kitchen-sink of `:test-run`. This shrinks `test_run.py` by 30%.

As a drive-by I eliminate a few more uses of `__file__` by replacing them with `layer_resource`, which in turn lets me get rid of some `par_style` incantations.

Reviewed By: justintrudell

Differential Revision: D22148464

fbshipit-source-id: 166ffd95b7128bcb4233bef6f11aee5c46a206c5