`--shadow-path` stashes shadowed originals in `/__fs_image__/shadowed`
Summary:
This second (and final) part of `--shadow-path` serves two purposes:
 - When the shadow is a wrapper of the original, it can access the original without needing to mess with mount namespaces to remove the shadow, etc.
 - This lets the container update the original via `librename_shadowed.so`.

You will notice that there's some handling around the fact that `/__fs_image__/shadowed` may not be on the same FS as the path being shadowed. This should only arise when we're trying to shadow something that is a mount. Specifically:
 - If `/usr/` is a host RO mount, then `cp --reflink=always /usr/... /__fs_image/shadow...` won't work because we're not on the same filesystem. So we use `cp --reflink=always`.
- On clean-up, trying to `cp` **back** into `/usr/` will necessarily fail because `/usr/bin/yum` is on a read-only FS.

For me, specifically, it came up when I was testing with a host build appliance like so:

```
buck run :fb_host_build_appliance-container -- --user=root --shadow-path /usr/bin/yum /__fs_image__/rpm/default-snapshot-for-installer/yum/yum/bin/yum --serve-rpm-snapshot /__fs_image__/rpm/default-snapshot-for-installer/yum/
```

This workflow allowed for much faster iteration on the `yum` / `dnf` wrappers (stacked) than rebuilding the real BA and using that.

However, outside of such "debug-only" cases, I feel like shadowing files inside mounts is a bit of a pathological case, so right now the container will exit with an error if the "shadow restore" operation fails.

Reviewed By: zeroxoneb

Differential Revision: D22215900

fbshipit-source-id: d210eac866633c581dda02f2a4fe015142841cf1