s/getfqdn/gethostname/ to allow container nesting
Summary:
There's no reason our containers shouldn't be able to nest inside other containers of ours. This is the only problem I found.

Getting `getfqdn` to work inside a Buck container resolves mucking with `/etc/hosts` or other kinds of DNS resolution, which suggests that it's the wrong tool for the job.

On the other hand, `buck run :foo-container -- --hostname "$HOSTNAME"` will result in a container that can successfully use the host's subvolumes.

Obviously, another alternative is to ditch the hostname checks entirely. We might do that later, but it requires a bit more deliberation.

Reviewed By: zeroxoneb

Differential Revision: D22318233

fbshipit-source-id: 55209948850d4b7112a7858f4463f5d186f7b1c3