Convert META_DIR,META_MOUNTS_DIR, and MOUNT_MARKER to be Path types
Summary:
As in summary.  Since these really are paths, let's type them as such.  Note: this doesn't go down the complete rabbit hole of moving *all* things that are paths to `Path`.  That probably should be done, but this just addreses the specific case for `META_DIR` and mounts.

Also, I did a drive by addition of `exists()` to the `Path` type for ease of use and for cool points.

Reviewed By: justintrudell

Differential Revision: D22532194

fbshipit-source-id: 0c61a95e0985313d6736c7e5c65e6f567d3c0097