Fix clean.sh and .flake8
Summary:
Noticed just doing `rm -rf buck-image-out/eden` is a bad idea as it'll hang if any Eden checkouts are left in that dir.

Cleaned it up a bit to just try to clear out any empty dirs and print a warning if checkouts were left over.

Also updated flake8 to ignore B950 as it's redundant with E501

Reviewed By: snarkmaster

Differential Revision: D22607077

fbshipit-source-id: 51c927ada98a130742dd659283aef0e8207216e5