Remove the only-for-tests-read-only-host-clone test-image and use the DEFAULT_BUILD_APPLIANCE instead
Summary:
The host-clone image layer is constructed entirely of host mounts and creates a strong dependency on the host system running the tests.   Instead this uses the DEFAULT_BUILD_APPLIANCE since we know that it already has the necerssary tools installed, and that requirement is unlikely to change considering that the BA is the *thing* that builds the layer.  As such, using it for testing components used during the build process makes sense.

Note also, that in the upcoming change where `systemd-nspawn` handles all of the mounts this method of constructing a layer entirely out of host mounts stops working.  This is due to the ordering of operations that `nspawn` uses, it fails to resolve the uid/gid of the requedted user because the binaries necessary for that (`getent`) and the passwd database is not available until *after* the bind mount is constructed which happens after.  It could be argued that this should be fixed within nspawn, however, *not* using a host-clone for operations within this tool is the correct (tm) thing to do anyway, since it ensures a more strict build and test environment and fully decouples it from the host.

NOTE:  This will need to be rebased after D22580992 lands and a new BA is built and released.

Reviewed By: justintrudell

Differential Revision: D22567985

fbshipit-source-id: d95cf2a9d4b41254e232e130c054256e18a3e60b