/meta -> /.meta
Summary:
We previously decided `.meta/` was a better name for this directory, so doing the migration now is beneficial, before we start filling the dir further and having downstream dependencies on it.

Once a new BA is built I should be ok to remove the TODOs in the changes below.

Reviewed By: snarkmaster

Differential Revision: D22669749

fbshipit-source-id: b538db6345bc8a896f4673a2b9c968a12d1d92e3