fs_image: allow installing the same version of an rpm twice
Summary:
Today, it is currently impossible to install the same version of and rpm twice,
`fs_image` will loudly complain if this is attempted.
There are some cases where this is unavoidable, as recently encountered due to
an upgrade of `kernel-headers` which is a version match for a synced `vmtest`
kernel with a test that requires `kernel-headers` to be installed into the VM.
`kernel-headers` is brought into the base vmtest image as a dependency of
`gcc`, which was recently introduced into the build appliance.

Technically this re-installs the rpm AFAICT, but is a much easier fix than
making it a proper no-op, and it should not really ever happen anyway - and
long term this interface should be explicit that it is allowed to re-install
the same version.

Reviewed By: snarkmaster

Differential Revision: D22822143

fbshipit-source-id: 082981c6eea7a7bd181d49acb2cacb869ebebfd6