fs_image: ignore selinux attrs when rendering subvol
Summary:
Previously, only the most common selinux attr value was removed, which was
fine, but this diff sets up image tests to completely ignore selinux
attributes.

This was prompted by me doing a kernel upgrade and suddenly seeing failing
tests. While this could be important to support in the future (for OSS at
least), at this time we shouldn't really be in the business of ensuring that
yum does-the-right-thing wrt selinux, unless and until selinux is a real
first-class citizen of image builds.

Reviewed By: zeroxoneb

Differential Revision: D22985946

fbshipit-source-id: 9c2175061f47a32b08d663d940e58ecf9ddd31b3