Re-sync md, add styling
Summary:
I'm going to officially point the team to using the md docs from now on, so am syncing some remaining changes from our wikis in the meantime to try to be up to date before doing so.

I also added a random temp logo of a moose until we get a real one, and had great fun with css.

Reviewed By: snarkmaster, vmagro

Differential Revision: D23011825

fbshipit-source-id: 20601c70cb731c89bf63805a948ea5cc2af4dc13