fs_image: shape - strongly typed bridge between bzl and python
Summary:
There exists a common need to convert some structured data at build time into
something that can be consumed at runtime.
Today, this is done by `echo`ing json into files with `buck_genrule`, but then
the schemas defined in python code need to carefully match the structure of
data exported by the buck macros.

This diff bridges that gap with a `shape` implementation. This allows a rule
author to define a shape of fields and their types that can then be validated
at build time, as well as codegen'ing Python loaders that are strongly
type(hinte)d, with automatic runtime type checking.

I expect there will be lots of bikeshedding necessary on the ergonomics of
this feature. This diff includes unit tests that demonstrate how a shape is
created in starlark macros, and how it is loaded by python code, which is
probably the most useful parts to review immediately, as the implementation may
have to change if there are changes needed to the user interface.

Reviewed By: justintrudell

Differential Revision: D22963480

fbshipit-source-id: 8ad08cfcea978b2a89dd099896355739a08d3e4f