Add no-op plumbing for "rpm version sets"
Summary:
A version set is implemented a repo path, namespaced per package manager. Right now, we only use the `rpm` sub-directory of the version set. Inside `rpm` is a collection of sibling targets -- one per available RPM. The output of such a target is the TAB-separated  ENVRA format that `yum_dnf_versionlock.py` expects. In the future, we may split `rpm` into multiple subdirectories if Buck parsing overhead proves problematic.

Version sets are RPM-agnostic in case we have to deal with similar version-selection problems for other package managers. There's nothing RPM-specific about the current plumbing. The rest of this text concerns only our RPM-centric usage of version sets.

We need to support a minimum of two RPM version-sets in practice:
 - "No version set" for the scenarios where "repo latest" is all the user needs.
 - "Policy-driven version set" -- if one repo mixes bleeding-edge RPMs with stable RPMs, we may want "production" builds to be constrained to avoid the bleeding-edge stuff. This is essentially papering over non-standard repo management practices.

In the "no version set" case, we want the build setup to work exactly as today.

When we do have a version set, we need a more elaborate dependency setup. Consider this scenario:
```
image.feature(
  name='rpm',
  features=[image.rpms_install('foo')],
)
image.layer(
  name='img',
  features=[':rpm', image.rpms_install('bar')],
)
```

Let's say that the "default version set" for our repo is `//my/ver_set` with shortname `my_ver_set`.

We want `:img` to depend on `//my/ver_set:foo` and `//my/ver_set:bar` so that whenever the version lock files change, the image gets rebuilt.

Therefore, `:rpm` has to depend on `//my/ver_set:foo`. However, `:rpm` does not know the version set of the layer including it. To deal with this, we generate multiple variants of `:rpm`:

- `:rpm_SIGIL` for "no version set", with no extra dependencies.
- `:rpm_SIGIL__rpm_verset__my_ver_set`, which depends on `//my/ver_set:foo`.

With this setup, `image.layer` simply has to know to depend on the appropriate variant of all of its "by target name" features, and it will get the right transitive dependencies.

I believe that this mess of split target could be fixed if Buck supported Bazel-style providers, but who knows when that might happen...

The current diff does not yet hook this version set data to be used in `rpm_action.py`, that's done later on the stack.

Reviewed By: zeroxoneb

Differential Revision: D22888690

fbshipit-source-id: 0f2d61fef368ccfe8e6f113dd15133170d079cc2